Create uml_model.Model as abstract class for UML models
This commit adds uml_model.Model as an abstract class for more UML models that can come after open sourcing the project. (e.g. class diagram, sequence diagram)

This commit makes it easier to create more UML TensorFlow models; future models can be made just by importing uml_model.py and implementing __init__ similarly to the one in usecase_model.Model.
